--- pkg/transport/kafka/configure_producer.go
@@ -1,58 +1,0 @@
-package kafka
-
-import (
-	"errors"
-	"fmt"
-	"os"
-	"strconv"
-
-	"github.com/confluentinc/confluent-kafka-go/kafka"
-)
-
-const (
-	envVarKafkaProducerID       = "KAFKA_PRODUCER_ID"
-	envVarKafkaBootstrapServers = "KAFKA_BOOTSTRAP_SERVERS"
-	envVarKafkaProducerTopic    = "KAFKA_PRODUCER_TOPIC"
-	envVarMessageSizeLimit      = "KAFKA_MESSAGE_SIZE_LIMIT_KB"
-)
-
-var (
-	errEnvVarNotFound     = errors.New("environment variable not found")
-	errEnvVarIllegalValue = errors.New("environment variable illegal value")
-)
-
-func getKafkaConfig() (*kafka.ConfigMap, string, int, error) {
-	id, found := os.LookupEnv(envVarKafkaProducerID)
-	if !found {
-		return nil, "", 0, fmt.Errorf("%w: %s", errEnvVarNotFound, envVarKafkaProducerID)
-	}
-
-	servers, found := os.LookupEnv(envVarKafkaBootstrapServers)
-	if !found {
-		return nil, "", 0, fmt.Errorf("%w: %s", errEnvVarNotFound, envVarKafkaProducerID)
-	}
-
-	topic, found := os.LookupEnv(envVarKafkaProducerTopic)
-	if !found {
-		return nil, "", 0, fmt.Errorf("%w: %s", errEnvVarNotFound, envVarKafkaProducerID)
-	}
-
-	sizeLimitString, found := os.LookupEnv(envVarMessageSizeLimit)
-	if !found {
-		return nil, "", 0, fmt.Errorf("%w: %s", errEnvVarNotFound, envVarKafkaProducerID)
-	}
-
-	sizeLimit, err := strconv.Atoi(sizeLimitString)
-	if err != nil {
-		return nil, "", 0, fmt.Errorf("%w: %s", errEnvVarIllegalValue, envVarMessageSizeLimit)
-	}
-
-	kafkaConfigMap := &kafka.ConfigMap{
-		"bootstrap.servers": servers,
-		"client.id":         id,
-		"acks":              "1",
-		"retries":           "0",
-	}
-
-	return kafkaConfigMap, topic, sizeLimit, nil
-}
